pyproject.toml
[project]
name = "vimeo-transcriber"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp>=2024.1.0",
    "openai-whisper",
    "torch",
    "anthropic>=0.28.0",
]

[project.scripts]
vimeo-transcriber = "vimeo_transcriber.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vimeo_transcriber"]
